+++ CHANGELOG.md
@@ -13,7 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Replace Base 64 buggy conversion functions with `uint8arrays`. 
+- Replace Base 64 buggy conversion functions with `uint8arrays`.
+
+### Removed
+
+- **Breaking**: Removed `equalByteArrays`, use `uint8arrays/equals` instead.
+  See changes in `eth-pm` example.
 
 ## [0.19.2] - 2022-03-21